application/Backend/requirements.txt
flask
bcrypt
transformers
torch
gunicorn
googletrans-py